Mejoras rutinarias

+++ src/Restaurant/Model/Bebida.java
@@ -3,7 +3,7 @@ package Restaurant.Model;
 public class Bebida extends Articulo{
     private double centilitros;
 
-    public Bebida(String nombre, String descripcion, double precio, double centilitros) {
+    public Bebida(String nombre, String descripcion, float precio, double centilitros) {
         super(nombre, descripcion, precio);
         this.centilitros = centilitros;
     }
